Contact_sales.feature has been modified by QA Mehmet Sen
