+++ src/test/java/runners/CukesRunnerTest.java
@@ -7,14 +7,14 @@ import org.junit.runner.RunWith;
 @RunWith(Cucumber.class)
 @CucumberOptions(
         plugin = {"pretty",
-                "html:target/reports/html/cucumber-droneFlightPlanner_reports.html",
+                "html:target/reports/html/cucumber-toqio.co_reports.html",
                 "json:target/reports/json/json-reports/cucumber.json",
                 "junit:target/reports/xml/xml-report/cucumber.xml",
                 "rerun:target/reports/rerun.txt"
         },
-        features = "src/test/resources/Features",
-        glue = "com/droneFlightPlanner/stepDefinitions",
-        tags = "@flightPlan"
+        features = "src/test/resources/features",
+        glue = "stepDefinitions",
+        tags = "@HP_01 or @CS_01"
 )
 public class CukesRunnerTest {
 }
