+++ src/test/java/stepDefinitions/Contact_sales_stepDef.java
@@ -4,37 +4,21 @@ import io.cucumber.datatable.DataTable;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
-import org.junit.Assert;
-import org.openqa.selenium.By;
-import org.openqa.selenium.Keys;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.ui.Select;
 import pages.CommonPage;
 import utilities.BrowserUtils;
 
-import java.util.Arrays;
-import java.util.List;
-
-import static stepDefinitions.Hooks.actions;
 import static stepDefinitions.Hooks.driver;
 
 public class Contact_sales_stepDef extends CommonPage {
 
     @When("user clicks the button Request Demo")
     public void userClicksTheButtonRequestDemo() {
-        actions.sendKeys(Keys.PAGE_DOWN).build().perform();
-
-        driver.switchTo().frame(contactSales().iframeElement);
-        contactSales().requestDemoButton.get(0).click();
+        contactSales().clickRequestDemoButton();
     }
 
     @Then("user verifies the URL of the new tab is {string}")
     public void userVerifiesTheURLOfTheNewTabIs(String url) {
-        BrowserUtils.switchToNewTab(driver);
-        BrowserUtils.waitForPageToLoad(25);
-        String currentUrl = driver.getCurrentUrl();
-        Assert.assertEquals("The current URL doesn't match the expected URL", url, currentUrl);
-        System.out.println("The Current URL is: " + currentUrl);
+        contactSales().verifyUrlOnTheNewTab(url);
     }
 
     @Then("user verifies the title Contact us now! appears")
@@ -44,38 +28,12 @@ public class Contact_sales_stepDef extends CommonPage {
 
     @Then("user verifies the Contact us inboxes and checkboxes are functional")
     public void userVerifiesTheContactUsInboxesAndCheckboxesAreFunctional() {
-        List<WebElement> elementsToCheck = Arrays.asList(
-                contactSales().firstnameInbox,
-                contactSales().lastnameInbox,
-                contactSales().emailInbox,
-                contactSales().phoneNumberInbox,
-                contactSales().companyNameInbox,
-                contactSales().country_RegionInbox,
-                contactSales().location_of_headquartersInbox,
-                contactSales().agreementCheckbox,
-                contactSales().subscriptionCheckbox
-        );
-
-        for (WebElement element : elementsToCheck) {
-            verifyElementFunctional(element);
-        }
-    }
-
-    private void verifyElementFunctional(WebElement element) {
-        Assert.assertTrue("Element is not displayed", element.isDisplayed());
-        Assert.assertTrue("Element is not enabled", element.isEnabled());
+        contactSales().verifyContactUsInboxesCheckboxes();
     }
 
     @And("user fills the First name inbox")
     public void userFillsTheFirstNameInbox() {
-        List<WebElement> iframes = driver.findElements(By.tagName("iframe"));
-
-        if (iframes.size() >= 8) {
-            driver.switchTo().frame(iframes.get(7));
-        } else {
-            System.out.println(iframes.size() + " iframes were found");
-        }
-        contactSales().firstnameInbox.sendKeys("Mehmet");
+        contactSales().fillFirstName();
     }
 
     @And("user fills the Last name inbox")
@@ -105,17 +63,12 @@ public class Contact_sales_stepDef extends CommonPage {
 
     @And("user selects a country from the Headquartered in inbox")
     public void userSelectsACountryFromTheHeadquarteredInInbox() {
-        BrowserUtils.waitAndClick(contactSales().location_of_headquartersInbox, 2);
-        Select select = new Select(contactSales().location_of_headquartersInbox);
-        select.selectByValue("Spain");
-//        actions.sendKeys(Keys.TAB).build().perform();
+        contactSales().selectsACountry();
     }
 
     @And("user fills the How did you hear about us? inbox")
     public void userFillsTheHowDidYouHearAboutUsInbox() {
-        contactSales().howDidYouHearAboutUsInbox.click();
-        Select select = new Select(contactSales().howDidYouHearAboutUsInbox);
-        select.selectByValue("Online advertising");
+        contactSales().fillHowDidYouHear();
     }
 
     @And("user checks the Agreement checkbox")
@@ -130,43 +83,26 @@ public class Contact_sales_stepDef extends CommonPage {
 
     @And("user clicks the button Send")
     public void userClicksTheButtonSend() {
-        actions.sendKeys(Keys.PAGE_DOWN).build().perform();
-        BrowserUtils.scrollToElement(driver, contactSales().sendButton);
-        BrowserUtils.waitForClickability(contactSales().sendButton);
-        contactSales().sendButton.click();
+        contactSales().clickSendButton();
     }
 
     @Then("user verifies the Error Messages on the required fields")
     public void userVerifiesTheErrorMessagesOnTheRequiredFields(DataTable dataTable) {
-        List<String> expectedMessages = dataTable.asList();
-        List<WebElement> errorElements = Arrays.asList(contactSales().errorMessage1, contactSales().errorMessage2);
-
-        for (String expectedMessage : expectedMessages) {
-            boolean messageFound = false;
-            for (WebElement errorElement : errorElements) {
-                if (errorElement.isDisplayed() && errorElement.getText().equals(expectedMessage)) {
-                    messageFound = true;
-                    break;
-                }
-            }
-            Assert.assertTrue("Error message not found: " + expectedMessage, messageFound);
-        }
-
-        System.out.println("All expected error messages are displayed and verified: " + expectedMessages);
+        contactSales().verifyErrorMessage(dataTable);
     }
 
     @And("user clicks the button Send on the Contact Us page")
     public void userClicksTheButtonSendOnTheContactUsPage() {
-        List<WebElement> iframes = driver.findElements(By.tagName("iframe"));
+        contactSales().clickSendButtonOnContactUs();
+    }
 
-        if (iframes.size() >= 8) {
-            driver.switchTo().frame(iframes.get(7));
-        } else {
-            System.out.println(iframes.size() + " iframes were found");
-        }
-        BrowserUtils.scrollToBottom();
-        BrowserUtils.scrollToElement(driver, contactSales().sendButton);
-        BrowserUtils.waitForClickability(contactSales().sendButton);
-        contactSales().sendButton.click();
+    @And("user switches to the new tab")
+    public void userSwitchesToTheNewTab() {
+        BrowserUtils.switchToNewTab(driver);
+    }
+
+    @And("user switches to the default tab")
+    public void userSwitchesToTheDefaultTab() {
+        BrowserUtils.switchToDefaultTab(driver);
     }
 }
